I2C: Track slave status
In situations where master repeatedly opens new transaction,
the I2C IRQ might trip up and declares that it's being selected,
but actually it's in the middle of a read transaction.
To prevent that, we implement a simple state machine
to keep track of the I2C slave status
(whether it's waiting for an address match/receiving/transmitting).

By doing that we also have an enum conflict witn i2c_rw_status,
so we change I2C_READ to I2C_RO and I2C_READ_WRITE to I2C_RW.